+++ src/AssemblyInfo.Html.Core.fs
@@ -5,13 +5,13 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FSharp.Data.Html.Core")>]
 [<assembly: AssemblyProductAttribute("FSharp.Data")>]
 [<assembly: AssemblyDescriptionAttribute("Library of F# type providers and data access tools")>]
-[<assembly: AssemblyVersionAttribute("6.2.0.0")>]
-[<assembly: AssemblyFileVersionAttribute("6.2.0.0")>]
+[<assembly: AssemblyVersionAttribute("6.5.0.0")>]
+[<assembly: AssemblyFileVersionAttribute("6.5.0.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "FSharp.Data.Html.Core"
     let [<Literal>] AssemblyProduct = "FSharp.Data"
     let [<Literal>] AssemblyDescription = "Library of F# type providers and data access tools"
-    let [<Literal>] AssemblyVersion = "6.2.0.0"
-    let [<Literal>] AssemblyFileVersion = "6.2.0.0"
+    let [<Literal>] AssemblyVersion = "6.5.0.0"
+    let [<Literal>] AssemblyFileVersion = "6.5.0.0"